[bugfix] l4xnat: fix farm config and farmguardian with pid 0

	modified:   usr/share/perl5/Zevenet/Farm/L4ipt/Action.pm
	modified:   usr/share/perl5/Zevenet/Farm/L4ipt/Backend.pm
	modified:   usr/share/perl5/Zevenet/Farm/L4ipt/Config.pm
	modified:   usr/share/perl5/Zevenet/FarmGuardian.pm

+++ usr/share/perl5/Zevenet/Farm/L4ipt/Action.pm
@@ -307,13 +307,14 @@ sub setL4NewFarmName    # ($farm_name,$new_farm_name)
 
 	my $farm       = &getL4FarmStruct( $farm_name );
 	my $fg_enabled = ( &getFarmGuardianConf( $$farm{ name } ) )[3];
-	my $fg_pid     = &getFarmGuardianPid( $farm_name );
+	my $fg_pid;
 
 	if ( $$farm{ status } eq 'up' )
 	{
 		if ( $fg_enabled eq 'true' )
 		{
-			kill 'STOP' => $fg_pid;
+			$fg_pid = &getFarmGuardianPid( $farm_name );
+			kill 'STOP' => $fg_pid if ( $fg_pid > 0 );
 		}
 	}
 
@@ -430,7 +431,7 @@ sub setL4NewFarmName    # ($farm_name,$new_farm_name)
 
 		if ( $fg_enabled eq 'true' )
 		{
-			if ( $0 !~ /farmguardian/ )
+			if ( $0 !~ /farmguardian/ && $fg_pid > 0 )
 			{
 				kill 'CONT' => $fg_pid;
 			}
